do not close and open menu again with one click #168, also refactor component
